Move param structures to dedicated file and add PreprocessingParameters
