+++ planning/apparent_safe_velocity_limiter/src/debug.cpp
@@ -14,7 +14,7 @@
 
 #include "apparent_safe_velocity_limiter/debug.hpp"
 
-#include "apparent_safe_velocity_limiter/forward_projection.hpp"
+#include "apparent_safe_velocity_limiter/parameters.hpp"
 #include "apparent_safe_velocity_limiter/types.hpp"
 
 #include <visualization_msgs/msg/marker_array.hpp>
